Extract UI components to class
